+++ packages/commerce-sdk-react/src/provider.tsx
@@ -58,6 +58,7 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
     apiClients?: ApiClients
     disableAuthInit?: boolean
     hybridAuthEnabled?: boolean
+    cookieDomain?: string
     pageDesignerParams?: PageDesignerParams
 }
 
@@ -157,6 +158,7 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         apiClients,
         disableAuthInit = false,
         hybridAuthEnabled = false,
+        cookieDomain,
         pageDesignerParams = {}
     } = props
 
@@ -183,7 +185,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
             passwordlessLoginCallbackURI,
             refreshTokenRegisteredCookieTTL,
             refreshTokenGuestCookieTTL,
-            hybridAuthEnabled
+            hybridAuthEnabled,
+            cookieDomain
         })
     }, [
         clientId,
@@ -205,7 +208,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         refreshTokenRegisteredCookieTTL,
         refreshTokenGuestCookieTTL,
         apiClients,
-        hybridAuthEnabled
+        hybridAuthEnabled,
+        cookieDomain
     ])
 
     const dwsid = auth.get(DWSID_COOKIE_NAME)